+++ typescript/src/hooks/akronWeightedLVRFeeHook.ts
@@ -6,7 +6,6 @@ import {
 } from '@/weighted/AkronWeightedMath';
 
 export type HookStateAkronWeightedLVRFee = {
-    lastBalances: bigint[];
     weights: bigint[];
     minimumSwapFeePercentage: bigint;
 };
@@ -32,19 +31,13 @@ export class AkronWeightedLVRFeeHook implements HookBase {
             params.swapKind === SwapKind.GivenIn
                 ? _computeSwapFeePercentageGivenExactIn(
                     params.balancesLiveScaled18[params.indexIn],
-                    hookState.lastBalances[params.indexIn],
                     hookState.weights[params.indexIn],
-                    params.balancesLiveScaled18[params.indexOut],
-                    hookState.lastBalances[params.indexOut],
                     hookState.weights[params.indexOut],
                     params.amountGivenScaled18,
                 )
                 : _computeSwapFeePercentageGivenExactOut(
-                    params.balancesLiveScaled18[params.indexIn],
-                    hookState.lastBalances[params.indexIn],
                     hookState.weights[params.indexIn],
                     params.balancesLiveScaled18[params.indexOut],
-                    hookState.lastBalances[params.indexOut],
                     hookState.weights[params.indexOut],
                     params.amountGivenScaled18,
                 );
