+++ typescript/src/vault/vault.ts
@@ -34,7 +34,7 @@ import { defaultHook } from '../hooks/constants';
 import { ExitFeeHook } from '../hooks/exitFeeHook';
 import { DirectionalFeeHook } from '../hooks/directionalFeeHook';
 import { StableSurgeHook } from '../hooks/stableSurgeHook';
-import { AkronWeightedLVRFeeHook } from '../hooks/akronWeightedLVRFeeHook';
+import { AkronWeightedLVRFeeHook } from '../hooks/akronWeightedLVRFeeHookCompex';
 
 const _MINIMUM_TRADE_AMOUNT = 1e6;
 // const _MINIMUM_WRAP_AMOUNT = 1e3;
